+++ docs/changes.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+Version 0.3.0
+=============
+
+Release TBD
+
 Version 0.2.0
 =============